+++ ChangeLog.md
@@ -1,3 +1,9 @@
+## Version 0.2.1.0
+
+* Added ToPretty instances for type level strings and numbers
+* Added ToPretty instances for types of kind Maybe and Bool
+* Added ToPretty instances for some standard number types (Word*, Int*)
+
 ## Version 0.2.0.0
 
 * Added indentation support